test: colocate recorder project helper
