+++ e2e/fake-audio/recorder-archive.test.ts
@@ -1,6 +1,6 @@
 import { expect, type Page, test } from "@playwright/test";
-import { createRecorderProject } from "../helpers";
 import {
+  createRecorderProject,
   enableInput,
   seekRecorderByPixels,
   waitForRecordingSamples,
