+++ e2e/fake-audio/recorder-clip-selection.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
-import { createRecorderProject } from "../helpers";
 import {
+  createRecorderProject,
   enableInput,
   seekRecorderByPixels,
   waitForRecordingSamples,
