+++ e2e/helpers.ts
@@ -34,17 +34,6 @@ export async function waitForAudioReady(page: Page): Promise<void> {
   await expect(page.getByTestId("play-pause-button")).toBeEnabled();
 }
 
-/** Create a recorder project from its index and wait for the recorder app. */
-export async function createRecorderProject(page: Page): Promise<void> {
-  await page.goto("/recorder");
-  await page.getByTestId("new-recorder-project-button").click();
-  await expect(page).toHaveURL(/\/recorder\/[^/]+$/);
-  await expect(page.getByTestId("recorder-project-name")).toBeVisible();
-}
-
-/** @deprecated Use clickNewProject instead */
-export const clickThroughStartup = clickNewProject;
-
 /**
  * Load an audio file via Settings dialog.
  * Opens Settings, uploads the given fixture WAV under the given file name,
